[Refactor] Html 클래스 추가 - html 리턴 함수 추가
