+++ students/28th/gayun/js/main.js
@@ -1,14 +1,8 @@
+import { Html } from './html.js';
 
 const $search = document.querySelector('.search');
 const $feeds = document.querySelector('.feeds');
-const $commentInput = document.querySelector('.comment__input--text');
-const $commentButton = document.querySelector('.comment__input--button');
-const $commentContainer = document.querySelector('.comments');
-const $comments = document.querySelector('.comment');
-const $feedButtons = document.querySelector('.feed__button');
-const $deleteButton = document.querySelector('.comment__delete-button');
 const $myButton = document.querySelector('.my');
-const $peopleWhoLike = document.querySelector('.people-who-like__comment');
 const $searchResultContainer = document.querySelector('.search__result--container');
 const $searchResultWrap = document.querySelector('.search__result--wrap');
 
@@ -27,6 +21,8 @@ fetchData()
 
 let displayCount = 0;
 
+const html = new Html();
+
 function addComment(e) {
     e.preventDefault();
     const target = e.target;
@@ -34,15 +30,7 @@ function addComment(e) {
     const $commentContainer = $commentInput.parentNode.previousSibling.previousSibling;
     const newCommentEl = document.createElement('ul');
     newCommentEl.classList.add('comment');
-    newCommentEl.innerHTML = `
-        <span class="comment_item">
-            <span class="comment__id">${myId}</span>
-            <span class="comment__text">${$commentInput.value}</span>
-        </span>
-        <button class="comment__delete-button">
-            <i class="fas fa-times"></i>
-        </button>
-    `
+    newCommentEl.innerHTML = html.writeComment(myId, $commentInput.value)
     $commentContainer.appendChild(newCommentEl);
     $commentInput.value = '';
 }
@@ -79,24 +67,7 @@ function searchId(e) {
     } else {
         $searchResultWrap.classList.add('hide');
     }
-    const filteredId = idArr.filter((id) => {
-        return id[1].includes(target.value);
-    })
-    const defaultImgUrl = 'images/default_profile-img.png';
-    const html = filteredId.map((id) => {
-        return `
-            <li class="search__result--item">
-                <img src="${id[0] ?? defaultImgUrl}" alt="" class="search__result--img">
-                <span class="search__result--text">
-                    <p class="search__result--title">${id[1]}</p>
-                    <p class="search__result--subtitle">${id[2]}</p>
-                </span>
-            </li>
-
-        `
-    }).join('');
-    $searchResultContainer.innerHTML = html;
-    console.log(html)
+    $searchResultContainer.innerHTML = html.searchResult(target.value, idArr)
 }
 
 function handleMyMenu(e) {
@@ -108,32 +79,11 @@ function handleMyMenu(e) {
     }
 }
 
-const skeletonHtml = `
-        <div class="skeleton__header">
-            <span class="skeleton__header--profileImg"></span>
-            <span class="skeleton__header--id"></span>
-        </div>
-        <div class="skeleton__img">
-            <img class="loading" src="images/loading.gif" alt="">
-        </div>
-        <div class="skeleton__comment">
-            <span class="skeleton__comment--id"></span>
-            <span class="skeleton__comment--text"></span>
-        </div>
-        <div class="skeleton__comment">
-            <span class="skeleton__comment--id"></span>
-            <span class="skeleton__comment--text" id="comment--text2"></span>
-        </div>
-        <div class="skeleton__comment--input"></div>
-`
+const skeletonHtml = html.addSkeleton();
 const $feedEnd = document.querySelector('.feed__end');
-console.log($feedEnd)
 const callback = (entry, observer) => {
     if(entry[0].isIntersecting && entry[0].intersectionRatio === 1) {
-        console.log(entry)
          loading();
-    } else {
-
     }
 }
 
@@ -176,68 +126,7 @@ function displayFeed(feedEl, index) {
 
     if(data) {
         feedEl.classList.add('feed');
-        const commentHtml = data.comment.map((com) => {
-            return `
-                <ul class="comment">
-                    <span class="comment__item">
-                        <span class="comment__id">${com.commentId}</span>
-                        <span class="comment__text">${com.commentText}</span>
-                    </span>
-                    <button class="comment__delete-button">
-                        <i class="fas fa-times"></i>
-                    </button>
-                </ul>`
-        }).join('');
-
-        feedEl.innerHTML = `
-                <div class="feed__header">
-                    <span class="feed__header--user">
-                        <span class="feed__header--img-container profile-container">
-                            <img class="feed__header--img profile-img" src="${data.profileImg}" alt="">
-                        </span>
-                        <span class="feed__header--id">${data.feedId}
-                        </span>
-                    </span>
-                    <button class="feed_header--button">
-                        <i class="fas fa-ellipsis-h"></i>
-                    </button>
-                </div>
-                <div class="feed__img-container">
-                    <img src="${data.imgContent}" alt="" class="feed-img">
-                </div>
-                <div class="feed__contents-wrap">
-                    <div class="feed__buttons">
-                        <span class="feed__buttons1">
-                            <button class="feed__button">
-                                <i class="far fa-heart"></i>
-                            </button>
-                            <button class="feed__button">
-                                <i class="far fa-comment"></i>
-                            </button>
-                            <button class="feed__button">
-                                <i class="far fa-share-square"></i>
-                            </button>
-                        </span>
-                        <span class="feed__buttons2">
-                            <button class="feed__button">
-                                <i class="far fa-bookmark"></i>
-                            </button>
-                        </span>
-                    </div>
-                    <div class="people-who-like">
-                        <img src="images/profile-img2.jpg" alt="" class="people-who-like__img">
-                        <p class="people-who-like__comment" data-num=${index}>${data.likesCount[0]}님 외 ${data.likesCount.length}명이 좋아합니다.</p>
-                    </div>
-                    <div class="feed__textContent">${data.textContent}</div>
-                    <li class="comments">
-                        ${commentHtml}
-                    </li>
-                </div>
-                <form class="comment__input">
-                    <input type="text" placeholder=" 댓글 달기..." name="" id="" class="comment__input--text">
-                    <button class="comment__input--button">게시</button>
-                </form>
-        `
+        feedEl.innerHTML = html.addFeed(data, index);
     } else {
         feedEl.classList.add('noFeedContainer');
         feedEl.innerHTML = noFeedHtml;
@@ -245,9 +134,4 @@ function displayFeed(feedEl, index) {
     }
 }
 
-const noFeedHtml = `
-    <div class="noFeed">
-        <i id="noFeed__icon" class="fas fa-exclamation-circle"></i>
-        <p class="noFeed__text">불러올 피드가 없습니다.</p>
-    </div>
-`
+const noFeedHtml = html.handleNoFeed();
