+++ students/28th/gayun/main.html
@@ -129,6 +129,6 @@
             </footer>
         </div>
     </main>
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
