+++ students/28th/gayun/style/main_right.css
@@ -140,7 +140,7 @@
 
 .recommend__item--button {
     font-weight: 700;
-    color: blue;
+    color: var(--blue);
 }
 
 .hide {
